+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # OSGL Tool Change Log
 
 1.18.0
+* `$.getProperty` issue with `Map` type object #166
 * `IO.read` stage - support ByteBuffer and byte[] #164
 * Add built-in TypeConverter between ByteBuffer and byte array #163
 * `IO` write API completeness #162